Move some files, cleanup tests
