+++ src/components/jamInputForm/JamInput.css
@@ -1,0 +1,34 @@
+.Input-Border {
+    width: 400px;
+    height: 40px;
+    border-radius: 4px;
+    background-color: #ffffff;
+    border: solid 1px #8da1c3;
+}
+
+.jamInput {
+    width: 363px;
+    height: 30px;
+    font-family: Roboto;
+    font-size: 14px;
+    line-height: 2.14;
+    text-align: left;
+    color: #2b3346;
+}
+
+.Submit-Button {
+    width: 200px;
+    height: 40px;
+    border-radius: 4px;
+    background-color: #d1d1e2;
+}
+
+.Submit-Button-Text {
+    width: 70px;
+    height: 19px;
+    font-family: Roboto;
+    font-size: 16px;
+    font-weight: 500;
+    text-align: center;
+    color: #ffffff;
+}
